bugfix: sed failure when detecting relay networks in swarm environment
In standalone deployment, docker sets just a single IP address on the eth0 so
the current entrypoint.sh works without a problem, however when overlay network
is used there are two IP addresses configured on container's eth0.
When that happens, underlying sed expression fails because it suddenly has
unescaped multiline string to process, which is not a valid sed expression.

In my lab, docker configured following IPs on the eth0: 10.0.0.6/24 and
10.0.0.7/32. Based on that, I think including broader range which covers both
   makes sense here - hence we simply ignore anything that is on further lines.

Should solve https://github.com/namshi/docker-smtp/issues/18